+++ src/main/java/com/vetclinic/demo/model/Service.java
@@ -42,7 +42,7 @@ public class Service {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
