+++ src/main/java/com/vetclinic/demo/model/request/ServiceRequest.java
@@ -17,7 +17,7 @@ public class ServiceRequest {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
